refactored diff as a standalone tool that checks the databases (and only them)
