+++ src/pump/_bitstream.py
@@ -1,5 +1,6 @@
+import os
 import logging
-from ._utils import read_json, time_method, serialize, deserialize, progress_bar, log_before_import, log_after_import, path_exists
+from ._utils import read_json, time_method, serialize, deserialize, progress_bar, log_before_import, log_after_import
 
 _logger = logging.getLogger("pump.bitstream")
 
@@ -51,7 +52,7 @@ class bitstreams:
 
     @staticmethod
     def bitstream_path(internal_id: str):
-        return f'{internal_id[:2]}/{internal_id[2:4]}/{internal_id[4:6]}/{internal_id}'
+        return os.path.join(internal_id[:2], internal_id[2:4], internal_id[4:6], internal_id)
 
     @property
     def imported(self):
@@ -147,11 +148,11 @@ class bitstreams:
         log_key = "bitstreams"
         log_before_import(log_key, expected)
 
-        # TODO(jm): fake bitstreams
-        TEST_DEV5 = "http://dev-5.pc" in env["backend"]["endpoint"]
-        if TEST_DEV5 and env["assetstore"] == "":
-            _logger.error(
-                'Location of assetstore folder is not defined but it should be checked!')
+        test_instance = env["backend"].get("testing", False)
+        path_assetstore = env["assetstore"]
+        if test_instance and path_assetstore == "":
+            _logger.critical(
+                'Location of assetstore dir is not defined but it should be checked!')
 
         for i, b in enumerate(progress_bar(self._bs)):
             b_id = b['bitstream_id']
@@ -206,9 +207,11 @@ class bitstreams:
                 'primaryBundle_id': None
             }
 
-            # TODO(jm): fake bitstreams
             path = self.bitstream_path(params['internal_id'])
-            if TEST_DEV5 and not path_exists(f'{env["assetstore"]}{path}'):
+            full_path = os.path.join(path_assetstore, path)
+            # NOTE: if it is the testing instance AND we do not have the bitstream
+            # use our testing one
+            if test_instance and not os.path.exists(full_path):
                 data['sizeBytes'] = 1748
                 data['checkSum'] = {
                     'checkSumAlgorithm': b['checksum_algorithm'],
