--- tools/localization/out.json
@@ -1,1 +1,0 @@
-{}
