+++ tools/repo_diff/diff_settings/tul.py
@@ -1,0 +1,19 @@
+settings = {
+    "db_dspace_7": {
+        "name": "dspace7",
+
+        "host": "localhost",
+        "user": "postgres",
+        "password": "dspace",
+        "port": 5432,
+    },
+
+    "db_dspace_6": {
+        "name": "dspace6",
+
+        "host": "localhost",
+        "user": "postgres",
+        "password": "dspace",
+        "port": 5432,
+    }
+}
